requirements.txt
Flask==3.0.3
gunicorn==22.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
openai==1.35.10
python-dotenv==1.0.1
pillow
